requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.3
Pillow==8.2.0
scipy==1.6.3
Werkzeug==1.0.1
